requirements.txt
pypdf
python-dotenv
chromadb
google-genai
